cc1: add float support

+++ lex.h
@@ -1,7 +1,6 @@
 extern unsigned token;
 extern unsigned line_num;
-/* value will change as we handle more type stuff */
-extern unsigned token_value;
+extern unsigned long token_value;
 extern char filename[16];
 
 extern void next_token(void);
